requirements.txt
attrs==26.1.0
blinker==1.9.0
click==8.3.2
colorama==0.4.6
flasgger==0.9.7.1
Flask==3.1.3
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mistune==3.2.0
packaging==26.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
SQLAlchemy==2.0.49
typing_extensions==4.15.0
Werkzeug==3.1.8
